refactor(application): extract invoice processing and replay flows
